implement T-implement-uuid-generation
