+++ packages/ui-shared/src/types/settings/LlmProviderCardProps.ts
@@ -1,0 +1,14 @@
+import type { LlmConfigData } from "./LlmConfigData";
+
+/**
+ * Props interface for LlmProviderCard component
+ */
+export interface LlmProviderCardProps {
+  api: LlmConfigData & {
+    id: string;
+    provider: "openai" | "anthropic";
+  };
+  onEdit: () => void;
+  onDelete: () => void;
+  className?: string;
+}
