+++ packages/ui-shared/src/types/settings/index.ts
@@ -32,6 +32,9 @@ export * from "./EnhancedNavigationListProps";
 export * from "./FontSizePreviewProps";
 export * from "./generalSettings";
 export * from "./InteractiveTabsProps";
+export * from "./LlmConfigData";
+export * from "./LlmConfigModalProps";
+export * from "./LlmProviderCardProps";
 export * from "./ModalFooterProps";
 export * from "./ModalHeaderProps";
 export * from "./ModelSelectProps";
