+++ planning/projects/P-update-api-keys-tab-to-llm-setup/epics/E-transform-api-keys-to-llm-setup/features/F-dynamic-api-management/tasks-done/20250804_134601-T-implement-uuid-generation.md
@@ -1,14 +1,15 @@
 ---
 kind: task
 id: T-implement-uuid-generation
+parent: F-dynamic-api-management
+status: done
 title: Implement UUID generation utility for API configurations
-status: open
 priority: high
 prerequisites: []
 created: "2025-08-04T13:36:09.082744"
-updated: "2025-08-04T13:36:09.082744"
+updated: "2025-08-04T13:40:04.653814"
 schema_version: "1.1"
-parent: F-dynamic-api-management
+worktree: null
 ---
 
 ## Context
@@ -44,3 +45,7 @@ The dynamic API management feature requires unique IDs for each API configuratio
 - Keep implementation simple and focused
 
 ### Log
+
+**2025-08-04T18:46:01.681400Z** - Implemented UUID generation utility for API configurations using crypto.randomUUID() with fallback to timestamp + random string. The utility follows existing codebase patterns and provides robust unique ID generation for the LLM Setup feature. Function is properly exported through barrel file and integrated into LlmSetupSection component to replace simple timestamp-based ID generation. All quality checks pass successfully.
+
+- filesChanged: ["apps/desktop/src/components/settings/llm-setup/generateId.ts", "apps/desktop/src/components/settings/llm-setup/index.ts", "apps/desktop/src/components/settings/LlmSetupSection.tsx"]
